+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.4.4] - 2023-02-17
 ### Added
 - Polka City URL validation
 ### Changed
@@ -863,7 +865,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0]
 
-[Unreleased]: https://github.com/adshares/adserver/compare/v2.4.3...develop
+[Unreleased]: https://github.com/adshares/adserver/compare/v2.4.4...develop
+[2.4.4]: https://github.com/adshares/adserver/compare/v2.4.3...v2.4.4
 [2.4.3]: https://github.com/adshares/adserver/compare/v2.4.2...v2.4.3
 [2.4.2]: https://github.com/adshares/adserver/compare/v2.4.1...v2.4.2
 [2.4.1]: https://github.com/adshares/adserver/compare/v2.4.0...v2.4.1